+++ meridian/analysis/analyzer.py
@@ -4177,14 +4177,14 @@ class Analyzer:
     """Computes the point-wise mean and credible intervals for the Hill curves.
 
     Args:
-      channel_type: Type of channel, either `media` or `rf`.
+      channel_type: Type of channel, either `media`, `rf`, or `organic_media`.
       confidence_level: Confidence level for `posterior` and `prior` credible
         intervals, represented as a value between zero and one.
 
     Returns:
       A DataFrame with data needed to plot the Hill curves, with columns:
 
-      *   `channel`: `media` or `rf` channel name.
+      *   `channel`: `media`, `rf`, or `organic_media` channel name.
       *   `media_units`: Media (for `media` channels) or average frequency (for
           `rf` channels) units.
       *   `distribution`: Indication of `posterior` or `prior` draw.
@@ -4193,7 +4193,12 @@ class Analyzer:
       *   `ci_lo`: Lower bound of the credible interval of the value of the Hill
           function.
       *   `mean`: Point-wise mean of the value of the Hill function per draw.
-      *   channel_type: Indication of a `media` or `rf` channel.
+      *   channel_type: Indication of a `media`, `rf`, or `organic_media`
+          channel.
+
+    Raises:
+      ValueError: If `channel_type` is not one of the recognized constants
+      `media`, `rf`, or `organic_media`.
     """
     if (
         channel_type == constants.MEDIA
@@ -4201,31 +4206,46 @@ class Analyzer:
     ):
       ec = constants.EC_M
       slope = constants.SLOPE_M
-      linspace = np.linspace(
-          0,
-          np.max(
-              np.array(self._meridian.media_tensors.media_scaled), axis=(0, 1)
-          ),
-          constants.HILL_NUM_STEPS,
-      )
       channels = self._meridian.input_data.media_channel.values
+      transformer = self._meridian.media_tensors.media_transformer
+      linspace_max_values = np.max(
+          np.array(self._meridian.media_tensors.media_scaled), axis=(0, 1)
+      )
     elif (
         channel_type == constants.RF
         and self._meridian.input_data.rf_channel is not None
     ):
       ec = constants.EC_RF
       slope = constants.SLOPE_RF
-      linspace = np.linspace(
-          0,
-          np.max(np.array(self._meridian.rf_tensors.frequency), axis=(0, 1)),
-          constants.HILL_NUM_STEPS,
-      )
       channels = self._meridian.input_data.rf_channel.values
+      transformer = None
+      linspace_max_values = np.max(
+          np.array(self._meridian.rf_tensors.frequency), axis=(0, 1)
+      )
+    elif (
+        channel_type == constants.ORGANIC_MEDIA
+        and self._meridian.input_data.organic_media_channel is not None
+    ):
+      ec = constants.EC_OM
+      slope = constants.SLOPE_OM
+      channels = self._meridian.input_data.organic_media_channel.values
+      transformer = (
+          self._meridian.organic_media_tensors.organic_media_transformer
+      )
+      linspace_max_values = np.max(
+          np.array(self._meridian.organic_media_tensors.organic_media_scaled),
+          axis=(0, 1),
+      )
     else:
       raise ValueError(
-          f"Unsupported channel type: {channel_type} or the"
-          " requested type of channels (`media` or `rf`) are not present."
+          f"Unsupported channel type: {channel_type} or the requested type of"
+          " channels (`media`, `rf`, or `organic_media`) are not present."
       )
+    linspace = np.linspace(
+        0,
+        linspace_max_values,
+        constants.HILL_NUM_STEPS,
+    )
     linspace_filler = np.linspace(0, 1, constants.HILL_NUM_STEPS)
     xr_dims = [
         constants.MEDIA_UNITS,
@@ -4281,13 +4301,10 @@ class Analyzer:
 
     # Fill media_units or frequency x-axis with the correct range.
     media_units_arr = []
-    if channel_type == constants.MEDIA:
-      media_transformers = transformers.MediaTransformer(
-          self._meridian.media_tensors.media, self._meridian.population
-      )
-      population_scaled_median_m = media_transformers.population_scaled_median_m
+    if transformer is not None:
+      population_scaled_median = transformer.population_scaled_median_m
       x_range_full_shape = linspace * tf.transpose(
-          population_scaled_median_m[:, np.newaxis]
+          population_scaled_median[:, np.newaxis]
       )
     else:
       x_range_full_shape = linspace
@@ -4308,8 +4325,68 @@ class Analyzer:
     df[constants.MEDIA_UNITS] = media_units_arr
     return df
 
+  def _get_channel_hill_histogram_dataframe(
+      self,
+      channel_type: str,
+      data_to_histogram: tf.Tensor,
+      channel_names: Sequence[str],
+      n_bins: int,
+  ) -> pd.DataFrame:
+    """Calculates hill histogram dataframe for a given channel type's values.
+
+    Args:
+      channel_type: The type of channel (e.g., 'rf', 'media', 'organic_media').
+      data_to_histogram: The 2D tensor (observations, channels). containing the
+        data whose distribution needs to be histogrammed for each channel.
+      channel_names: The names corresponding to the channels in
+        data_to_histogram.
+      n_bins: The number of bins for the histogram.
+
+    Returns:
+      A Pandas DataFrame containing the calculated histogram data for all
+      channels of the given type. Returns an empty DataFrame if no valid
+      data is found for any channel.
+    """
+    channels_data = {
+        constants.CHANNEL: [],
+        constants.CHANNEL_TYPE: [],
+        constants.SCALED_COUNT_HISTOGRAM: [],
+        constants.COUNT_HISTOGRAM: [],
+        constants.START_INTERVAL_HISTOGRAM: [],
+        constants.END_INTERVAL_HISTOGRAM: [],
+    }
+
+    for i, channel_name in enumerate(channel_names):
+      channel_data_np = data_to_histogram[:, i].numpy()
+      channel_data_np = channel_data_np[~np.isnan(channel_data_np)]
+      if channel_data_np.size == 0:
+        continue
+
+      counts_per_bucket, buckets = np.histogram(
+          channel_data_np, bins=n_bins, density=True
+      )
+      max_counts = (
+          np.max(counts_per_bucket) if np.max(counts_per_bucket) > 0 else 1.0
+      )
+
+      num_buckets = len(counts_per_bucket)
+      channels_data[constants.CHANNEL].extend([channel_name] * num_buckets)
+      channels_data[constants.CHANNEL_TYPE].extend([channel_type] * num_buckets)
+      channels_data[constants.SCALED_COUNT_HISTOGRAM].extend(
+          counts_per_bucket / max_counts
+      )
+      channels_data[constants.COUNT_HISTOGRAM].extend(counts_per_bucket)
+      channels_data[constants.START_INTERVAL_HISTOGRAM].extend(buckets[:-1])
+      channels_data[constants.END_INTERVAL_HISTOGRAM].extend(buckets[1:])
+
+    return pd.DataFrame(channels_data)
+
   def _get_hill_histogram_dataframe(self, n_bins: int) -> pd.DataFrame:
-    """Returns the bucketed media_units counts per each `media` or `rf` channel.
+    """Calculates histogram data for a given channel type's values.
+
+      Computes histogram data for the distribution of media units (for media or
+      organic media channels) or frequency (for RF channels) across
+      observations.
 
     Args:
       n_bins: Number of equal-width bins to include in the histogram for the
@@ -4335,73 +4412,64 @@ class Analyzer:
     """
     n_geos = self._meridian.n_geos
     n_media_times = self._meridian.n_media_times
-    n_rf_channels = self._meridian.n_rf_channels
-    n_media_channels = self._meridian.n_media_channels
 
-    (
-        channels,
-        scaled_count,
-        channel_type_arr,
-        start_interval_histogram,
-        end_interval_histogram,
-        count,
-    ) = ([], [], [], [], [], [])
+    df_list = []
 
     # RF.
     if self._meridian.input_data.rf_channel is not None:
-      frequency = (
-          self._meridian.rf_tensors.frequency
-      )  # Shape: (n_geos, n_media_times, n_channels).
-      reshaped_frequency = tf.reshape(
-          frequency, (n_geos * n_media_times, n_rf_channels)
-      )
-      for i, channel in enumerate(self._meridian.input_data.rf_channel.values):
-        # Bucketize the histogram data for RF channels.
-        counts_per_bucket, buckets = np.histogram(
-            reshaped_frequency[:, i], bins=n_bins, density=True
+      frequency = self._meridian.rf_tensors.frequency
+      if frequency is not None:
+        reshaped_frequency = tf.reshape(
+            frequency, (n_geos * n_media_times, self._meridian.n_rf_channels)
         )
-        channels.extend([channel] * len(counts_per_bucket))
-        channel_type_arr.extend([constants.RF] * len(counts_per_bucket))
-        scaled_count.extend(counts_per_bucket / max(counts_per_bucket))
-        count.extend(counts_per_bucket)
-        start_interval_histogram.extend(buckets[:-1])
-        end_interval_histogram.extend(buckets[1:])
+        rf_hist_data = self._get_channel_hill_histogram_dataframe(
+            channel_type=constants.RF,
+            data_to_histogram=reshaped_frequency,
+            channel_names=self._meridian.input_data.rf_channel.values,
+            n_bins=n_bins,
+        )
+        df_list.append(pd.DataFrame(rf_hist_data))
 
     # Media.
     if self._meridian.input_data.media_channel is not None:
-      transformer = transformers.MediaTransformer(
-          self._meridian.media_tensors.media, self._meridian.population
-      )
-      scaled = (
-          self._meridian.media_tensors.media_scaled
-      )  # Shape: (n_geos, n_media_times, n_channels)
-      population_scaled_median = transformer.population_scaled_median_m
-      scaled_media_units = scaled * population_scaled_median
-      reshaped_scaled_media_units = tf.reshape(
-          scaled_media_units, (n_geos * n_media_times, n_media_channels)
-      )
-      for i, channel in enumerate(
-          self._meridian.input_data.media_channel.values
-      ):
-        # Bucketize the histogram data for media channels.
-        counts_per_bucket, buckets = np.histogram(
-            reshaped_scaled_media_units[:, i], bins=n_bins, density=True
+      transformer = self._meridian.media_tensors.media_transformer
+      scaled = self._meridian.media_tensors.media_scaled
+      if transformer is not None and scaled is not None:
+        population_scaled_median = transformer.population_scaled_median_m
+        scaled_media_units = scaled * population_scaled_median
+        reshaped_scaled_media_units = tf.reshape(
+            scaled_media_units,
+            (n_geos * n_media_times, self._meridian.n_media_channels),
         )
-        channel_type_arr.extend([constants.MEDIA] * len(counts_per_bucket))
-        channels.extend([channel] * (len(counts_per_bucket)))
-        scaled_count.extend(counts_per_bucket / max(counts_per_bucket))
-        count.extend(counts_per_bucket)
-        start_interval_histogram.extend(buckets[:-1])
-        end_interval_histogram.extend(buckets[1:])
+        media_hist_data = self._get_channel_hill_histogram_dataframe(
+            channel_type=constants.MEDIA,
+            data_to_histogram=reshaped_scaled_media_units,
+            channel_names=self._meridian.input_data.media_channel.values,
+            n_bins=n_bins,
+        )
+        df_list.append(pd.DataFrame(media_hist_data))
+    # Organic media.
+    if self._meridian.input_data.organic_media_channel is not None:
+      transformer_om = (
+          self._meridian.organic_media_tensors.organic_media_transformer
+      )
+      scaled_om = self._meridian.organic_media_tensors.organic_media_scaled
+      if transformer_om is not None and scaled_om is not None:
+        population_scaled_median_om = transformer_om.population_scaled_median_m
+        scaled_organic_media_units = scaled_om * population_scaled_median_om
+        reshaped_scaled_organic_media_units = tf.reshape(
+            scaled_organic_media_units,
+            (n_geos * n_media_times, self._meridian.n_organic_media_channels),
+        )
+        organic_media_hist_data = self._get_channel_hill_histogram_dataframe(
+            channel_type=constants.ORGANIC_MEDIA,
+            data_to_histogram=reshaped_scaled_organic_media_units,
+            channel_names=self._meridian.input_data.organic_media_channel.values,
+            n_bins=n_bins,
+        )
+        df_list.append(pd.DataFrame(organic_media_hist_data))
 
-    return pd.DataFrame({
-        constants.CHANNEL: channels,
-        constants.CHANNEL_TYPE: channel_type_arr,
-        constants.SCALED_COUNT_HISTOGRAM: scaled_count,
-        constants.COUNT_HISTOGRAM: count,
-        constants.START_INTERVAL_HISTOGRAM: start_interval_histogram,
-        constants.END_INTERVAL_HISTOGRAM: end_interval_histogram,
-    })
+    return pd.concat(df_list, ignore_index=True)
 
   def hill_curves(
       self,
@@ -4449,17 +4517,16 @@ class Analyzer:
       )
 
     final_dfs = [pd.DataFrame()]
-    if self._meridian.n_media_channels > 0:
-      hill_df_media = self._get_hill_curves_dataframe(
-          constants.MEDIA, confidence_level
-      )
-      final_dfs.append(hill_df_media)
-
-    if self._meridian.n_rf_channels > 0:
-      hill_df_rf = self._get_hill_curves_dataframe(
-          constants.RF, confidence_level
-      )
-      final_dfs.append(hill_df_rf)
+    for n_channels, channel_type in [
+        (self._meridian.n_media_channels, constants.MEDIA),
+        (self._meridian.n_rf_channels, constants.RF),
+        (self._meridian.n_organic_media_channels, constants.ORGANIC_MEDIA),
+    ]:
+      if n_channels > 0:
+        hill_df = self._get_hill_curves_dataframe(
+            channel_type, confidence_level
+        )
+        final_dfs.append(hill_df)
 
     final_dfs.append(self._get_hill_histogram_dataframe(n_bins=n_bins))
     return pd.concat(final_dfs)